+++ okitweb/static/okit/model/js/okit_model.js
@@ -444,11 +444,11 @@ class OkitArtifact {
     ** Default name generation
      */
     generateDefaultName(count = 0) {
-        return this.getNamePrefix()
+        // return this.getNamePrefix()
         // const today = new Date();
         // const pad = (n) => ("0" + n).slice(-2)
         // return `${this.getNamePrefix()}-${today.getFullYear()}${pad(today.getMonth() + 1)}${pad(today.getDate())}-${pad(today.getHours())}${pad(today.getMinutes())}${pad(today.getSeconds())}`;
-        // return this.getNamePrefix() + ('000' + count).slice(-3);
+        return this.getNamePrefix() + ('000' + count).slice(-3);
     }
 
     getNamePrefix() {